Deteccion Facial Arduino/requirements.txt
absl-py==1.3.0
attrs==22.1.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
contourpy==1.0.6
cycler==0.11.0
flatbuffers==22.12.6
fonttools==4.38.0
future==0.18.2
imutils==0.5.4
iso8601==1.1.0
kiwisolver==1.4.4
matplotlib==3.6.2
mediapipe==0.9.0.1
numpy==1.23.5
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
protobuf==3.20.3
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
wincertstore==0.2
